requirements.txt
annotated-types==0.6.0
anyio==3.7.1
Authlib==1.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.4
fastapi==0.104.0
h11==0.14.0
httptools==0.6.1
idna==3.4
Pillow==10.1.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
validators==0.22.0
watchfiles==0.21.0
weaviate-client==3.24.2
websockets==12.0
